requirements.txt
matplotlib==3.4.2
pyyaml==5.4.1
opencv-python==4.5.1.48
easydict==1.9
tensorboardx==2.1
tensorboard==2.13.0
lpips==0.1.3
imgaug==0.4.0
tqdm==4.60.0
imageio==2.9.0
click
requests
pyspng
ninja
imageio-ffmpeg==0.4.3
psutil
omegaconf
pytorch_fid
onnxruntime==1.13.1
fvcore
protobuf==3.20.*
numpy==1.23.5
